MacOSX Bundle stuff.

+++ mkMacDist.sh
@@ -1,7 +1,11 @@
 #!/bin/bash
 # Create a distributable installable package
 
-BUNDLE=$1
+if [ $# -gt 1 ]; then
+    BUNDLE=$1
+else
+    BUNDLE=scidavis.app
+fi
 
 MAC_DIST_DIR=$BUNDLE/Contents/MacOS
 RES_DIR=$BUNDLE/Contents/Resources
@@ -16,13 +20,16 @@ rewrite_dylibs()
     echo "rewrite_dylibs $target"
     otool -L $target|grep opt/local|cut -f1 -d' '|while read dylib; do
         # avoid infinite loops
-        if [ "${dylib##*/}" != "${target##*/}" ]; then 
+        if [ "${dylib##*/}" != "${target##*/}" -a ! -f $MAC_DIST_DIR/${dylib##*/} ]; then 
             cp -f $dylib $MAC_DIST_DIR
             chmod u+rw $MAC_DIST_DIR/${dylib##*/}
             rewrite_dylibs $MAC_DIST_DIR/${dylib##*/}
             echo "install_name_tool -change $dylib @executable_path/${dylib##*/} $target"
         fi
+        echo install_name_tool -change $dylib @executable_path/${dylib##*/} $target
         install_name_tool -change $dylib @executable_path/${dylib##*/} $target
+        echo install_name_tool -id @executable_path/${dylib##*/} $target
+        install_name_tool -id @executable_path/${dylib##*/} $target
     done
     otool -L $target|grep usr/local|cut -f1 -d' '|while read dylib; do
         # avoid infinite loops
@@ -31,35 +38,32 @@ rewrite_dylibs()
             chmod u+rw $MAC_DIST_DIR/${dylib##*/}
             rewrite_dylibs $MAC_DIST_DIR/${dylib##*/}
         fi
+        echo install_name_tool -change $dylib @executable_path/${dylib##*/} $target
         install_name_tool -change $dylib @executable_path/${dylib##*/} $target
     done
 }
 
+mkdir -p $MAC_DIST_DIR
 cp scidavis/scidavis $MAC_DIST_DIR
-rewrite_dylibs $MAC_DIST_DIR/scidavis
-
 chmod u+w $MAC_DIST_DIR/*
 
-# Generic resources required for Qt
-#cp -rf /opt/local/libexec/qt4/Library/Frameworks/QtGui.framework/Resources/qt_menu.nib $RES_DIR
 
-# python resources
+## python resources
 mkdir -p $RES_DIR/lib
 cp -rf /opt/local/Library/Frameworks/Python.framework/Versions/Current/lib/python* $RES_DIR/lib
 
 # python resources contain some dynamic libraries that need rewriting
 find $RES_DIR/lib -name "*.so" -print | while read soname; do
-    otool -L $soname|grep /opt/local|cut -f1 -d' '|while read oldName; do
-        install_name_tool -change $oldName @executable_path/${oldName##*/} $soname
-        done
+    rewrite_dylibs $soname
 done
 
 # copy translation files
 cp -rf scidavis/translations $MAC_DIST_DIR
 
 # copy icon, and create mainfest
+mkdir -p $RES_DIR
 cp -f scidavis/icons/scidavis.icns $RES_DIR
-cat >scidavis/scidavis.app/Contents/Info.plist <<EOF
+cat >$BUNDLE/Contents/Info.plist <<EOF
 <?xml version="1.0" encoding="UTF-8"?>
 <!DOCTYPE plist PUBLIC "-//Apple Computer//DTD PLIST 1.0//EN"
  "http://www.apple.com/DTDs/PropertyList-1.0.dtd">
@@ -87,6 +91,8 @@ cat >scidavis/scidavis.app/Contents/Info.plist <<EOF
 </dict>
 </plist>
 EOF
+
+macdeployqt $BUNDLE
 
 codesign -s "Developer ID Application" --deep --force $BUNDLE
 if [ $? -ne 0 ]; then
